+++ feature/home/src/test/kotlin/com/yapp/home/util/AlarmDateTimeFormatterTest.kt
@@ -1,4 +1,4 @@
-package com.yapp.domain.formatter
+package com.yapp.home.util
 
 import com.yapp.domain.model.Alarm
 import com.yapp.domain.model.AlarmDay
